Move WebAuthn adapter to separate file
